Removed Base64 conversion of the value & messages are been organized into packeges
