+++ src/main/scala/server/messages/ErrorMessages.scala
@@ -3,6 +3,7 @@ package server.messages
 /**
   * Created by matteobortolazzo on 02/05/2016.
   */
-trait ErrorMessages extends ActorbaseMessage{}
-
-case class InvalidQueryMessage() extends ErrorMessages
+object ErrorMessages {
+  trait ErrorMessages extends ActorbaseMessage {}
+  case class InvalidQueryMessage() extends ErrorMessages
+}
