+++ src/main/scala/server/messages/query/PermissionMessages.scala
@@ -1,0 +1,10 @@
+package server.messages.query
+
+/**
+  * Created by matteobortolazzo on 04/05/2016.
+  */
+object PermissionMessages {
+  trait NoPermissionMessage
+  trait ReadMessage extends NoPermissionMessage
+  trait ReadWriteMessage extends ReadMessage
+}
