+++ src/main/scala/server/messages/query/RowMessages.scala
@@ -1,0 +1,17 @@
+package server.messages.query
+
+import server.messages.ActorbaseMessage
+import server.messages.query.PermissionMessages.{ReadMessage, ReadWriteMessage}
+
+/**
+  * Created by matteobortolazzo on 02/05/2016.
+  */
+object RowMessages {
+  trait RowMessage extends ActorbaseMessage
+  case class InsertRowMessage(key: String, value: String) extends RowMessage with ReadWriteMessage
+  case class UpdateRowMessage(key: String, value: String) extends RowMessage with ReadWriteMessage
+  case class RemoveRowMessage(key: String) extends RowMessage with ReadWriteMessage
+  case class FindRowMessage(key: String) extends RowMessage with ReadMessage
+  case class ListKeysMessage() extends RowMessage with ReadMessage  // Message from storemanger to storefinder
+  case class StorefinderRowMessage(mapName: String, rowMessage: RowMessage) extends ActorbaseMessage
+}
