+++ src/main/scala/server/util/Parser.scala
@@ -4,6 +4,10 @@ import java.util
 
 import akka.event.LoggingAdapter
 import server.messages._
+import server.messages.query.DatabaseMessages._
+import server.messages.query.MapMessages._
+import server.messages.query.RowMessages._
+import ErrorMessages.InvalidQueryMessage
 
 import scala.util.matching.Regex
